Refactor and setup log handling more cleanly
